+++ common/src/main/java/com/yanny/ali/plugin/mods/cultural_delights/Plugin.java
@@ -1,0 +1,20 @@
+package com.yanny.ali.plugin.mods.cultural_delights;
+
+import com.yanny.ali.api.AliEntrypoint;
+import com.yanny.ali.api.IPlugin;
+import com.yanny.ali.api.IServerRegistry;
+import com.yanny.ali.plugin.mods.PluginUtils;
+
+@AliEntrypoint
+public class Plugin implements IPlugin {
+    @Override
+    public String getModId() {
+        return "villagersplus";
+    }
+
+    @Override
+    public void registerServer(IServerRegistry registry) {
+        PluginUtils.registerItemListing(registry, EmeraldToItemOffer.class);
+        PluginUtils.registerItemListingCollector(registry, EmeraldToItemOffer.class);
+    }
+}